requirements.txt
asgiref==3.4.1
blessings==1.7
boto3==1.20.6
botocore==1.23.6
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
curtsies==0.3.10
cwcwidth==0.1.5
Django==3.2.9
django-cors-headers==3.10.0
django-extensions==3.1.5
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
dnspython==2.1.0
drf-yasg==1.20.0
greenlet==1.1.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
packaging==21.2
Pillow==8.4.0
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
pyxdg==0.27
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.2.4
typing_extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
